Image Annotation Tool
LabelConverter is an image annotation tool designed for object and pose detection projects. It allows users to annotate images and directly convert label coordinates into text format. This simplifies the labeling process and enhances efficiency for researchers and developers in computer vision, making it an essential tool for machine learning workflows!